requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
cssselect==1.0.3
Django==2.1.4
fake-useragent==0.1.11
idna==2.8
lxml==4.2.5
parse==1.9.0
psycopg2-binary==2.7.6.1
pyee==5.0.0
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
requests-html==0.9.0
six==1.12.0
tqdm==4.28.1
urllib3==1.24.1
w3lib==1.19.0
watson-developer-cloud==2.5.1
websocket-client==0.48.0
websockets==7.0
